+++ inst/doc/miceFast-intro.R
@@ -2,7 +2,9 @@
 knitr::opts_chunk$set(echo = FALSE)
 
 ## ----echo=TRUE,message=FALSE,warning=FALSE-------------------------------
-library(miceFast);library(data.table);library(magrittr)
+library(miceFast);library(data.table);library(magrittr);library(mice);library(car);library(dplyr);library(ggplot2)
+
+## ----echo=TRUE-----------------------------------------------------------
 set.seed(123456)
 
 ## ----eval=FALSE,echo=TRUE------------------------------------------------
@@ -10,6 +12,180 @@ set.seed(123456)
 
 ## ----eval=FALSE,echo=TRUE------------------------------------------------
 #  system.file("extdata","images",package = "miceFast")
+
+## ----echo=TRUE-----------------------------------------------------------
+# airquality dataset with additional variables
+data(air_miss)
+
+## ----echo=TRUE-----------------------------------------------------------
+# VIF - values bigger than 10 (around) suggest that there might be a collinearity problem.
+# VIF is high for Solar.R and x_character which is obvious - x_character is a factor version of numeric Solar.R
+air_miss[,.(VIF(.SD,posit_y='Ozone',
+                   posit_x=c('Solar.R',
+                             'Wind',
+                             'Temp',
+                             'x_character',
+                             'Day',
+                             'weights',
+                             'groups')))] 
+                             
+# IMPUTATIONS
+# Imputations with a grouping option (models are separately assessed for each group) 
+# taking into account provided weights
+air_miss[,Solar_R_imp := fill_NA_N(x=.SD,
+                                   model="lm_bayes",
+                                   posit_y='Solar.R',
+                                   posit_x=c('Wind','Temp','Intercept'),
+                                   w=.SD[['weights']],
+                                   times=100),by=.(groups)] %>%
+# Imputations - discrete variable
+  .[,x_character_imp := fill_NA(x=.SD,
+                                model="lda",
+                                posit_y='x_character',
+                                posit_x=c('Wind','Temp','groups'))] %>%
+# logreg was used because almost log-normal distribution of Ozone
+# imputations around mean
+.[,Ozone_imp1 := fill_NA(x=.SD, 
+                         model="lm_bayes",
+                         posit_y='Ozone',
+                         posit_x=c('Intercept'),
+                         logreg=TRUE)] %>% 
+# imputations using positions - Intercept, Temp
+.[,Ozone_imp2 := fill_NA(x=.SD, 
+                         model="lm_bayes",
+                         posit_y=1,
+                         posit_x=c(4,6),
+                         logreg=TRUE)] %>% 
+# model with a factor independent variable 
+# multiple imputations (average of x30 imputations) 
+# with a factor independent variable, weights and logreg options
+  .[,Ozone_imp3 := fill_NA_N(x=.SD, 
+                             model="lm_noise",
+                             posit_y='Ozone',
+                             posit_x=c('Intercept','x_character_imp','Wind','Temp'),
+                             w=.SD[['weights']],
+                             logreg=TRUE,
+                             times=30)] %>% 
+  .[,Ozone_imp4 := fill_NA_N(x=.SD, 
+                             model="lm_bayes",
+                             posit_y='Ozone',
+                             posit_x=c('Intercept','x_character_imp','Wind','Temp'),
+                             w=.SD[['weights']],
+                             logreg=TRUE,
+                             times=30)] %>% 
+  .[,Ozone_imp5 := fill_NA(x=.SD, 
+                           model="lm_pred",
+                           posit_y='Ozone',
+                           posit_x=c('Intercept','x_character_imp','Wind','Temp'),
+                           w=.SD[['weights']],
+                           logreg=TRUE),.(groups)] %>%
+                           
+# Average of a few methods
+  .[,Ozone_imp_mix := apply(.SD,1,mean),.SDcols=Ozone_imp1:Ozone_imp5] %>% 
+  
+# Protecting against collinearity or low number of observations - across small groups
+# Be carful when using a data.table grouping option 
+# because of lack of protection against collinearity or low number of observations. 
+# There could be used a tryCatch(fill_NA(...),error=function(e) return(...))
+
+  .[,Ozone_chac_imp := tryCatch(fill_NA(x=.SD, 
+                                 model="lda",
+                                 posit_y='Ozone_chac',
+                                 posit_x=c('Intercept','Month','Day','Temp','x_character_imp'),
+                                 w=.SD[['weights']]),
+                                error=function(e) .SD[['Ozone_chac']]),.(groups)] 
+
+## ----echo=TRUE-----------------------------------------------------------
+# VIF - values bigger than 10 (around) suggest that there might be a collinearity problem.
+# VIF is high for Solar.R and x_character which is obvious - x_character is a factor version of numeric Solar.R
+air_miss %>% do(vifs=VIF(.,posit_y='Ozone',
+                           posit_x=c('Solar.R',
+                                     'Wind',
+                                     'Temp',
+                                     'x_character',
+                                     'Day',
+                                     'weights',
+                                     'groups'))) %>% unlist()
+                             
+# IMPUTATIONS
+air_miss = air_miss %>% 
+# Imputations with a grouping option (models are separately assessed for each group) 
+# taking into account provided weights
+group_by(groups) %>% 
+do(mutate(.,Solar_R_imp = fill_NA(x=.,
+                                  model="lm_pred",
+                                  posit_y='Solar.R',
+                                  posit_x=c('Wind','Temp','Intercept'),
+                                  w=.[['weights']]))) %>%
+ungroup() %>%
+# Imputations - discrete variable
+mutate(x_character_imp = fill_NA(x=.,
+                                 model="lda",
+                                 posit_y='x_character',
+                                 posit_x=c('Wind','Temp'))) %>%
+# logreg was used because almost log-normal distribution of Ozone
+# imputations around mean
+mutate(Ozone_imp1 = fill_NA(x=., 
+                            model="lm_bayes",
+                            posit_y='Ozone',
+                            posit_x=c('Intercept'),
+                            logreg=TRUE)) %>% 
+# imputations using positions - Intercept, Temp
+mutate(Ozone_imp2 = fill_NA(x=., 
+                            model="lm_bayes",
+                            posit_y=1,
+                            posit_x=c(4,6),
+                            logreg=TRUE)) %>% 
+# multiple imputations (average of x30 imputations) 
+# with a factor independent variable, weights and logreg options
+mutate(Ozone_imp3 = fill_NA_N(x=., 
+                              model="lm_noise",
+                              posit_y='Ozone',
+                              posit_x=c('Intercept','x_character_imp','Wind','Temp'),
+                              w=.[['weights']],
+                              logreg=TRUE,
+                              times=30)) %>% 
+mutate(Ozone_imp4 = fill_NA_N(x=., 
+                              model="lm_bayes",
+                              posit_y='Ozone',
+                              posit_x=c('Intercept','x_character_imp','Wind','Temp'),
+                              w=.[['weights']],
+                              logreg=TRUE,
+                              times=30)) %>% 
+group_by(groups) %>%
+do(mutate(.,Ozone_imp5 = fill_NA(x=., 
+                                 model="lm_pred",
+                                 posit_y='Ozone',
+                                 posit_x=c('Intercept','x_character_imp','Wind','Temp'),
+                                 w=.[['weights']],
+                                 logreg=TRUE))) %>%
+ungroup() %>%
+# Average of a few methods
+mutate(Ozone_imp_mix = rowMeans(select(.,starts_with("Ozone_imp")))) %>%
+
+# Protecting against collinearity or low number of observations - across small groups
+# Be carful when using a data.table grouping option 
+# because of lack of protection against collinearity or low number of observations. 
+# There could be used a tryCatch(fill_NA(...),error=function(e) return(...))
+group_by(groups) %>%
+do(mutate(.,Ozone_chac_imp = tryCatch(fill_NA(x=., 
+                                       model="lda",
+                                       posit_y='Ozone_chac',
+                                       posit_x=c('Intercept','Month','Day','Temp','x_character_imp'),
+                                       w=.[['weights']]),
+                                      error=function(e) .[['Ozone_chac']]))) %>%
+ungroup()
+
+## ----echo=TRUE-----------------------------------------------------------
+# Distribution of imputations vs Distribution of initial data
+air_miss$Ozone_NA = ifelse(is.na(air_miss$Ozone),'imputations','complete')
+air_miss = as.data.table(air_miss)
+
+air_miss[,c('Ozone','Ozone_imp_mix','Ozone_NA')] %>% 
+melt(id=c('Ozone_NA'),measure=c('Ozone','Ozone_imp_mix')) %>% 
+.[(((Ozone_NA=='imputations') & (variable=='Ozone_imp_mix'))|((Ozone_NA=='complete') & (variable=='Ozone'))),] %>% 
+ggplot2::ggplot(.,ggplot2::aes(x=value,fill=variable)) + 
+ggplot2::geom_density() + ggplot2::facet_wrap(Ozone_NA ~.,scales='free')
 
 ## ----echo=TRUE-----------------------------------------------------------
 #install.packages("mice")
@@ -108,17 +284,18 @@ mtcars_mat = model.matrix.lm(~.,mtcars,na.action="na.pass")
 ## ----echo=TRUE-----------------------------------------------------------
 airquality2 = airquality
 airquality2$Temp2 = airquality2$Temp**2
-#install.packages("car")
-#car::vif(lm(Ozone ~ ., data=airquality2))
+airquality2$Month = factor(airquality2$Month)
 
-airquality2_mat = as.matrix(airquality2)
-model = new(miceFast)
-model$set_data(airquality2_mat)
-as.vector(model$vifs(1,c(2,3,4,5,6,7)))
+car::vif(lm(Ozone ~ ., data=airquality2))
 
+## ----echo=TRUE-----------------------------------------------------------
 data_DT = data.table(airquality2)
-# VIF for variables at 1,3,4 positions - you include a y position to consider its NA values
-as.vector(data_DT[,.(vifs=VIF(x=as.matrix(.SD),
+data_DT[,.(vifs=VIF(x=.SD,
+                posit_y='Ozone',
+                posit_x=c('Solar.R','Wind','Temp','Month','Day','Temp2'),correct=FALSE))][['vifs']]
+
+data_DT[,.(vifs=VIF(x=.SD,
                 posit_y=1,
-                posit_x=c(2,3,4,5,6,7)))][['vifs']])
+                posit_x=c(2,3,4,5,6,7),correct=TRUE))][['vifs']]
+